backend/requirements.txt

Flask

Flask-Cors

Flask-JWT-Extended

Flask-PyMongo

google-api-python-client

google-auth

google-auth-oauthlib

flask==2.0.1

flask-cors==3.0.10

flask-pymongo==2.3.0

flask-jwt-extended==4.3.1

python-dotenv==0.19.0

werkzeug==2.0.1

pyjwt==2.3.0

pymongo==3.12.0

google-auth==2.3.3

gunicorn==20.1.0